Refactor the other get_gold into screen_helpers
